Move helper code to test_gemm_streamk_util.hpp
