--- test/ck_tile/gemm_streamk/smoke_tests/f16_rcr_compv3_128x128x32_2x2x1_32x32x16_NonPersistent.cpp
@@ -1,12 +1,0 @@
-// Copyright © Advanced Micro Devices, Inc., or its affiliates.
-// SPDX-License-Identifier:  MIT
-
-#include "../test_gemm_streamk_common_includes.hpp"
-
-
-#define TEST_SUITE_PARAMS f16_rcr_compv3_128x128x32_2x2x1_32x32x16_NonPersistent
-#define TEST_SUITE_NAME MAKE_TEST_SUITE_NAME(TEST_SUITE_PARAMS)
-
-DECLARE_STREAM_K_TEST(TEST_SUITE_NAME, TEST_SUITE_PARAMS);
-
-#include "../test_gemm_streamk_cases.inc"
