+++ test/ck_tile/gemm_streamk/test_gemm_streamk_types_fp16.hpp
@@ -5,7 +5,6 @@
 
 #include "test_gemm_streamk_types.hpp"
 
-// clang-format off
 template<typename M_MacroTile,
          typename N_MacroTile,
          typename K_MacroTile,
@@ -19,9 +18,10 @@ template<typename M_MacroTile,
          typename Persistent>
 struct F16Layouts
 {
+    // clang-format off
     // For CDNA, we support [A, B, Acc, C] = [f16, f16, f32, f16] and [f16, f16, f32, f32]:
     using f16_f16_f32_f16 = Layouts<F16, F16, F32, F16, M_MacroTile, N_MacroTile, K_MacroTile, M_Warps, N_Warps, K_Warps, M_MmaTile, N_MmaTile, K_MmaTile, PipelineType, Persistent>;
-    using f16_f16_f32_f32 = Layouts<F16, F16, F32, F16, M_MacroTile, N_MacroTile, K_MacroTile, M_Warps, N_Warps, K_Warps, M_MmaTile, N_MmaTile, K_MmaTile, PipelineType, Persistent>;
+    using f16_f16_f32_f32 = Layouts<F16, F16, F32, F32, M_MacroTile, N_MacroTile, K_MacroTile, M_Warps, N_Warps, K_Warps, M_MmaTile, N_MmaTile, K_MmaTile, PipelineType, Persistent>;
     using rrr = detail::combine_t<typename f16_f16_f32_f16::rrr, typename f16_f16_f32_f32::rrr>;
     using rrc = detail::combine_t<typename f16_f16_f32_f16::rrc, typename f16_f16_f32_f32::rrc>;
     using rcr = detail::combine_t<typename f16_f16_f32_f16::rcr, typename f16_f16_f32_f32::rcr>;
@@ -30,12 +30,14 @@ struct F16Layouts
     using crc = detail::combine_t<typename f16_f16_f32_f16::crc, typename f16_f16_f32_f32::crc>;
     using ccr = detail::combine_t<typename f16_f16_f32_f16::ccr, typename f16_f16_f32_f32::ccr>;
     using ccc = detail::combine_t<typename f16_f16_f32_f16::ccc, typename f16_f16_f32_f32::ccc>;
+    // clang-format on
 };
 
 template<typename PipelineType,
          typename Persistent>
 struct F16Set
 {
+    // clang-format off
     // 32x32x16
     // 2x2x1
     using f16_128x128x32_2x2x1_32x32x16 =  F16Layouts<I128, I128,  I32, I2, I2, I1, I32, I32, I16, PipelineType, Persistent>;
@@ -107,9 +109,10 @@ struct F16Set
     using f16_128x256x64_1x4x1_16x16x16 =  F16Layouts<I128, I256,  I64, I1, I4, I1, I16, I16, I16, PipelineType, Persistent>;
     using f16_256x256x32_1x4x1_16x16x16 =  F16Layouts<I256, I256,  I32, I1, I4, I1, I16, I16, I16, PipelineType, Persistent>;
     using f16_256x256x64_1x4x1_16x16x16 =  F16Layouts<I256, I256,  I64, I1, I4, I1, I16, I16, I16, PipelineType, Persistent>;
+    // clang-format on
 };
 
-
+// clang-format off
 // mem
 // 32x32x16
 // 2x2x1
@@ -135,3 +138,4 @@ using f16_crc_compv3_128x128x32_2x2x1_32x32x16_NonPersistent = typename F16Set<C
 using f16_ccr_compv3_128x128x32_2x2x1_32x32x16_NonPersistent = typename F16Set<CompV3, NonPersistent>::f16_128x128x32_2x2x1_32x32x16::ccr;
 using f16_ccc_compv3_128x128x32_2x2x1_32x32x16_NonPersistent = typename F16Set<CompV3, NonPersistent>::f16_128x128x32_2x2x1_32x32x16::ccc;
 
+// clang-format on
